08/requirements.txt
docopt==0.6.2
importlib-metadata==2.1.1
Pillow==8.2.0
xortool==1.0.1
zipp==3.4.1
